list05/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.6
sqlparse==0.4.3
psycopg2-binary